refactor(types): extract board task types from team types
Move the board task type block out of src/shared/types/team.ts into a new src/shared/types/teamBoardTask.ts and re-export it from team.ts.

The moved region is origin/main team.ts lines 198-420 (TaskRef through BoardTaskLogStreamSummary, 223 lines) and is byte-identical in the new module, verified by diffing it against `git show origin/main:src/shared/types/team.ts | sed -n '198,420p'`. The remainder of team.ts is likewise byte-identical to the original minus that region and minus the old `@main/types` import line, verified line by line.

The block contains no runtime values at all - every declaration in it is an interface or a type alias, so `export type *` is sufficient and the emitted JavaScript is unchanged. The only runtime value in team.ts, DEFAULT_TOOL_APPROVAL_SETTINGS, stays where it is.

The `import type { EnhancedChunk } from '@main/types'` on old line 3 moves to the new module, because both of its uses (BoardTaskExactLogDetail and BoardTaskLogSegment) travel with the block. team.ts gains `import type { TaskRef } from './teamBoardTask'` because eleven interfaces that stay behind still carry TaskRef fields. BoardTaskExactLogSummaryBase was module-private and stays module-private - it gets no export.

Import and export order in the team.ts header is load-bearing for simple-import-sort: the new type import sits between './notifications' and './teamProvisioning', and `export type * from './teamBoardTask'` sits before `export type * from './teamProvisioning'`.

No call site changes. All 94 files that reference the moved symbols reach them either through the @shared/types barrel (index.ts:28 re-exports team.ts) or directly from @shared/types/team, and both paths keep working through the re-export. Inside src/shared only api.ts:115 and schedule.ts:8 import from './team', neither for a symbol whose defining file changed. No test file changed.

team.ts 1958 -> 1735 lines against a frozen cap of 1958 (223 lines of headroom); teamBoardTask.ts is 225 lines, well under the 800-line limit for new files, so it needs no baseline entry. scripts/ci/source-file-size-baseline.json is not modified.
